refa: extract common methods into base class
