+++ lib/bundler/audit/presenter/default.rb
@@ -1,25 +1,9 @@
+require 'bundler/audit/presenter/base'
+
 module Bundler
   module Audit
     module Presenter
-      class Default
-        attr_reader :options
-        attr_reader :shell
-
-        def initialize(shell, options)
-          @shell = shell
-          @options = options
-          @warnings = []
-          @advisory_bundles = []
-        end
-
-        def push_warning(message)
-          @warnings.push(message)
-        end
-
-        def push_advisory(advisory)
-          @advisory_bundles.push(advisory)
-        end
-
+      class Default < Base
         def print_report
           @warnings.each do |warning|
             print_warning warning
@@ -36,15 +20,7 @@ module Bundler
           end
         end
 
-        def exit_code
-          problematic? ? 1 : 0
-        end
-
         protected
-
-        def problematic?
-          @warnings.any? || @advisory_bundles.any?
-        end
 
         def say(message = '', color = nil)
           color = nil unless $stdout.tty?
